% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apidemo-package.R
\docType{package}
\name{apidemo-package}
\alias{apidemo}
\alias{apidemo-package}
\title{apidemo: Plumber-as-Package Demonstration}
\description{
The goal of apidemo is to...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/apidemo}
  \item \url{https://jonthegeek.github.io/apidemo/}
  \item Report bugs at \url{https://github.com/jonthegeek/apidemo/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
